AspectJ
git-svn-id: https://faracvs.cs.uni-magdeburg.de/svn/tthuem-FeatureIDE/trunk@733 272fd0a3-3bb5-487d-978a-b927a3a9602e

+++ plugins/de.ovgu.featureide.core/src/de/ovgu/featureide/core/internal/BuilderMarkerHandler.java
@@ -104,10 +104,12 @@ public class BuilderMarkerHandler implements IBuilderMarkerHandler {
 	 *      int)
 	 */
 	public void deleteBuilderMarkers(IResource resource, int depth) {
-		try {
-			resource.deleteMarkers(BUILDER_MARKER, false, depth);
-		} catch (CoreException e) {
-			CorePlugin.getDefault().logError(e);
+		if (resource != null && resource.exists()) {
+			try {
+				resource.deleteMarkers(BUILDER_MARKER, false, depth);
+			} catch (CoreException e) {
+				CorePlugin.getDefault().logError(e);
+			}
 		}
 	}
